Renamed echo_log_method to default_log_method
Changed name of trait and file.

Included GPL-3 License
